Port Devise stuff from thredded_create_app
