+++ app/views/devise/passwords/edit.html.erb
@@ -1,0 +1,24 @@
+<h2><%= t(".change_your_password") %></h2>
+
+<%= simple_form_for(resource, as: resource_name, url: password_path(resource_name), html: { method: :put }) do |f| %>
+  <%= f.error_notification %>
+
+  <%= f.input :reset_password_token, as: :hidden %>
+  <%= f.full_error :reset_password_token %>
+
+  <div class="form-inputs">
+    <%= f.input :password,
+                label: t(".new_password"),
+                required: true,
+                autofocus: true,
+                hint: (t('devise.shared.minimum_password_length', count: @minimum_password_length) if @minimum_password_length),
+                input_html: { autocomplete: "new-password" } %>
+    <%= f.input :password_confirmation, label: t(".confirm_new_password"), required: true %>
+  </div>
+
+  <div class="form-actions">
+    <%= f.button :submit, t(".change_my_password") %>
+  </div>
+<% end %>
+
+<%= render "devise/shared/links" %>
