+++ app/views/devise/passwords/new.html.erb
@@ -1,0 +1,18 @@
+<h2><%= t(".forgot_your_password") %></h2>
+
+<%= simple_form_for(resource, as: resource_name, url: password_path(resource_name), html: { method: :post }) do |f| %>
+  <%= f.error_notification %>
+
+  <div class="form-inputs">
+    <%= f.input :email,
+                required: true,
+                autofocus: true,
+                input_html: { autocomplete: "email" } %>
+  </div>
+
+  <div class="form-actions">
+    <%= f.button :submit, t(".send_me_reset_password_instructions") %>
+  </div>
+<% end %>
+
+<%= render "devise/shared/links" %>
