+++ app/views/devise/registrations/edit.html.erb
@@ -1,4 +1,4 @@
-<h2><%= t(".title", resource: resource_name.to_s.humanize) %></h2>
+<h2><%= t(".title", resource: resource.model_name.human) %></h2>
 
 <%= simple_form_for(resource, as: resource_name, url: registration_path(resource_name), html: { method: :put }) do |f| %>
   <%= f.error_notification %>
@@ -26,11 +26,15 @@
       <p><%= t(".currently_waiting_confirmation_for_email", email: resource.unconfirmed_email) %></p>
     <% end %>
 
-    <%= f.input :password, autocomplete: "off", hint: t(".leave_blank_if_you_don_t_want_to_change_it"), required: false %>
-    <%= f.input :password_confirmation, required: false %>
+    <%= f.input :password,
+                hint: t(".leave_blank_if_you_don_t_want_to_change_it"),
+                required: false,
+                input_html: { autocomplete: "new-password" } %>
+    <%= f.input :current_password,
+                hint: t(".we_need_your_current_password_to_confirm_your_changes"),
+                required: true,
+                input_html: { autocomplete: "current-password" } %>
 
-    <hr />
-    <%= f.input :current_password, hint: t(".we_need_your_current_password_to_confirm_your_changes"), required: true %>
   </div>
 
   <div class="form-actions">
