+++ app/views/devise/registrations/new.html.erb
@@ -1,13 +1,18 @@
 <h2><%= t(".sign_up") %></h2>
 
-<%= simple_form_for(resource, as: resource_name, url: registration_path(resource_name)) do |f| %>
+<%= simple_form_for(resource, as: resource_name, url: registration_path(resource_name), html: {autocomplete: 'off'}) do |f| %>
   <%= f.error_notification %>
 
   <div class="form-inputs">
-  <%= f.input :display_name, required: true, autofocus: true %>
-    <%= f.input :email, required: true %>
-    <%= f.input :password, required: true, hint: (t('devise.shared.minimum_password_length', count: @minimum_password_length) if @minimum_password_length) %>
-    <%= f.input :password_confirmation, required: true %>
+    <%= f.input :display_name, required: true, autofocus: true %>
+    <%= f.input :email,
+                required: true,
+                autofocus: true ,
+                input_html: { autocomplete: "email" }%>
+    <%= f.input :password,
+                required: true,
+                hint: (t('devise.shared.minimum_password_length', count: @minimum_password_length) if @minimum_password_length),
+                input_html: { autocomplete: "new-password" } %>
   </div>
 
   <%= recaptcha_tags %>
