+++ app/views/devise/unlocks/new.html.erb
@@ -1,0 +1,19 @@
+<h2><%= t(".resend_unlock_instructions") %></h2>
+
+<%= simple_form_for(resource, as: resource_name, url: unlock_path(resource_name), html: { method: :post }) do |f| %>
+  <%= f.error_notification %>
+  <%= f.full_error :unlock_token %>
+
+  <div class="form-inputs">
+    <%= f.input :email,
+                required: true,
+                autofocus: true,
+                input_html: { autocomplete: "email" } %>
+  </div>
+
+  <div class="form-actions">
+    <%= f.button :submit, t(".resend_unlock_instructions") %>
+  </div>
+<% end %>
+
+<%= render "devise/shared/links" %>
