+++ config/routes.rb
@@ -5,7 +5,20 @@ Rails.application.routes.draw do
     end
   end
 
-  devise_for :users, controllers: { registrations: 'registrations' }
+  devise_for :users,
+             skip: %i[sessions],
+             controllers: {
+                 registrations: 'users/registrations',
+                 sessions: 'users/sessions',
+             },
+             path_names: { sign_up: 'register' }
+  devise_scope :user do
+    get 'sign-in', to: 'users/sessions#new', as: :new_user_session
+    post 'sign-in', to: 'users/sessions#create', as: :user_session
+    match 'sign-out', to: 'users/sessions#destroy', as: :destroy_user_session,
+          via: Devise.mappings[:user].sign_out_via
+  end
+
   resources :users, only: [:index, :show]
 
   resources :private_topics, only: [:new], controller: 'thredded/private_topics'
